New: fg-diag-all  Perftesting tool for live monitoring of FortiGate metrics.
Many of the metrics one may want to track doing perf testing or other testing
are still only available via the CLI.  Such as diag npu commands.  This program
can be used to query the FG for various metrics and output them to STDOUT and/or
an ouput file in json, json-pretty, csv, tsv or cacti for consumption by other
tools (such as Cacti).  A log file may also be specified in which a more human
readable output is sent everytime the program runs.  Generally would run this
program via a tool like Cacti, with cron or via another tool that continually
executes it on a determined interfal.  Documenation not yet complete.